refactor: move redux hooks to separate files
