+++ app/src/dashboardHooks.ts
@@ -1,0 +1,6 @@
+import { useAppSelector } from './storeHooks'
+import { selectDashboard } from './dashboardSlice'
+
+export function useDashboard() {
+  return useAppSelector(selectDashboard)
+}
