+++ app/src/logbookHooks.ts
@@ -1,0 +1,6 @@
+import { useAppSelector } from './storeHooks'
+import { selectLogbook } from './logbookSlice'
+
+export function useLogbook() {
+  return useAppSelector(selectLogbook)
+}
