+++ app/src/rosterHooks.ts
@@ -1,0 +1,6 @@
+import { useAppSelector } from './storeHooks'
+import { selectRoster } from './rosterSlice'
+
+export function useRoster() {
+  return useAppSelector(selectRoster)
+}
